requirements.txt
labjack-ljm
dearpygui